+++ packages/eslint-config/src/rules/import-rules.ts
@@ -1,3 +1,5 @@
+import { Linter } from 'eslint';
+
 // Import plugin configuration
 export const IMPORT_SETTINGS = {
   'import/resolver': {
@@ -16,7 +18,7 @@ export const IMPORT_SETTINGS = {
   'import/external-module-folders': ['node_modules', 'dist']
 };
 
-export const IMPORT_RULES = {
+export const IMPORT_RULES: Linter.RulesRecord = {
   'import/order': [
     'warn',
     {