+++ packages/eslint-config/src/rules/index.ts
@@ -1,0 +1,5 @@
+export * from './general-rules';
+export * from './import-rules';
+export * from './prettier-rules';
+export * from './tsdoc-rules';
+export * from './typescript-rules';